+++ public/r/data-table.json
@@ -193,6 +193,18 @@
       "content": "// Core table components\nexport { DataTableRoot } from \"./data-table-root\"\nexport type { DataTableConfig } from \"./data-table-root\"\nexport {\n  DataTableProvider,\n  useDataTable,\n  DataTableContext,\n} from \"./data-table-context\"\nexport { DataTable } from \"./data-table\"\nexport { DataTableErrorBoundary } from \"./data-table-error-boundary\"\nexport type { DataTableErrorBoundaryProps } from \"./data-table-error-boundary\"\n\n// Regular table structure (Header, Body, EmptyBody, Skeleton, Loading) - consolidated for easy copy/paste\nexport {\n  DataTableHeader,\n  DataTableBody,\n  DataTableEmptyBody,\n  DataTableSkeleton,\n  DataTableLoading,\n} from \"./data-table-structure\"\nexport type {\n  ScrollEvent,\n  DataTableHeaderProps,\n  DataTableBodyProps,\n  DataTableEmptyBodyProps,\n  DataTableSkeletonProps,\n  DataTableLoadingProps,\n} from \"./data-table-structure\"\n\n// Types\nexport type { DataTableContextState } from \"./data-table-context\"\nexport type { DataTableContainerProps } from \"./data-table\"\n",
       "type": "registry:lib",
       "target": "components/niko-table/core/index.tsx"
+    },
+    {
+      "path": "src/components/niko-table/components/index.tsx",
+      "content": "// Core Column Header components (context-aware wrappers)\nexport {\n  DataTableColumnHeader,\n  DataTableColumnHeaderRoot,\n  useColumnHeaderContext,\n} from \"./data-table-column-header\"\nexport { DataTableColumnTitle } from \"./data-table-column-title\"\nexport { DataTableColumnActions } from \"./data-table-column-actions\"\nexport { DataTableColumnFilter } from \"./data-table-column-filter\"\n\n// Core Toolbar\nexport { DataTableToolbarSection } from \"./data-table-toolbar-section\"\nexport type { DataTableToolbarSectionProps } from \"./data-table-toolbar-section\"\n\n// Core Empty state composition components\nexport {\n  DataTableEmptyIcon,\n  DataTableEmptyMessage,\n  DataTableEmptyFilteredMessage,\n  DataTableEmptyActions,\n  DataTableEmptyTitle,\n  DataTableEmptyDescription,\n} from \"./data-table-empty-state\"\n",
+      "type": "registry:lib",
+      "target": "components/niko-table/components/index.tsx"
+    },
+    {
+      "path": "src/components/niko-table/filters/index.tsx",
+      "content": "// Core column-level filter components\nexport { TableColumnTitle } from \"./table-column-title\"\nexport { TableColumnActions } from \"./table-column-actions\"\n",
+      "type": "registry:lib",
+      "target": "components/niko-table/filters/index.tsx"
     }
   ],
   "type": "registry:block"
