+++ src/components/niko-table/components/index.tsx
@@ -1,4 +1,4 @@
-// Column Header components (context-aware wrappers)
+// Core Column Header components (context-aware wrappers)
 export {
   DataTableColumnHeader,
   DataTableColumnHeaderRoot,
@@ -7,45 +7,12 @@ export {
 export { DataTableColumnTitle } from "./data-table-column-title"
 export { DataTableColumnActions } from "./data-table-column-actions"
 export { DataTableColumnFilter } from "./data-table-column-filter"
-export { DataTableColumnFilterTrigger } from "./data-table-column-filter-trigger"
-export {
-  DataTableColumnSortMenu,
-  DataTableColumnSortOptions,
-} from "./data-table-column-sort"
-export {
-  DataTableColumnHideOptions,
-  DataTableColumnHideMenu,
-} from "./data-table-column-hide"
-export {
-  DataTableColumnPinOptions,
-  DataTableColumnPinMenu,
-} from "./data-table-column-pin"
-export {
-  DataTableColumnFacetedFilterOptions,
-  DataTableColumnFacetedFilterMenu,
-} from "./data-table-column-faceted-filter"
-export {
-  DataTableColumnSliderFilterOptions,
-  DataTableColumnSliderFilterMenu,
-} from "./data-table-column-slider-filter-options"
-export {
-  DataTableColumnDateFilterOptions,
-  DataTableColumnDateFilterMenu,
-} from "./data-table-column-date-filter-options"
+
+// Core Toolbar
 export { DataTableToolbarSection } from "./data-table-toolbar-section"
 export type { DataTableToolbarSectionProps } from "./data-table-toolbar-section"
-export {
-  DataTableAside,
-  DataTableAsideTrigger,
-  DataTableAsideContent,
-  DataTableAsideHeader,
-  DataTableAsideTitle,
-  DataTableAsideDescription,
-  DataTableAsideClose,
-} from "./data-table-aside"
-export { DataTableSelectionBar } from "./data-table-selection-bar"
 
-// Empty state composition components
+// Core Empty state composition components
 export {
   DataTableEmptyIcon,
   DataTableEmptyMessage,
@@ -54,41 +21,3 @@ export {
   DataTableEmptyTitle,
   DataTableEmptyDescription,
 } from "./data-table-empty-state"
-
-// Context-aware filter components (previously in actions/)
-export { DataTableClearFilter } from "./data-table-clear-filter"
-export { DataTableViewMenu } from "./data-table-view-menu"
-export { DataTableSearchFilter } from "./data-table-search-filter"
-export { DataTableFacetedFilter } from "./data-table-faceted-filter"
-export { DataTableSliderFilter } from "./data-table-slider-filter"
-export { DataTableDateFilter } from "./data-table-date-filter"
-export { DataTableSortMenu } from "./data-table-sort-menu"
-export { DataTableFilterMenu } from "./data-table-filter-menu"
-export { DataTableInlineFilter } from "./data-table-inline-filter"
-export { DataTablePagination } from "./data-table-pagination"
-export { DataTableExportButton } from "./data-table-export-button"
-export type { DataTableExportButtonProps } from "./data-table-export-button"
-
-// Row DnD components (context-aware wrappers)
-export {
-  DataTableRowDndProvider,
-  DataTableDraggableRow,
-  DataTableRowDragHandle,
-} from "./data-table-row-dnd"
-export type {
-  DataTableRowDndProviderProps,
-  DataTableDraggableRowProps,
-  DataTableRowDragHandleProps,
-} from "./data-table-row-dnd"
-
-// Column DnD components (context-aware wrappers)
-export {
-  DataTableColumnDndProvider,
-  DataTableDraggableHeader,
-  DataTableDragAlongCell,
-} from "./data-table-column-dnd"
-export type {
-  DataTableColumnDndProviderProps,
-  DataTableDraggableHeaderProps,
-  DataTableDragAlongCellProps,
-} from "./data-table-column-dnd"
