+++ src/components/niko-table/filters/index.tsx
@@ -1,72 +1,3 @@
-// Core filter components
-export { TableSearchFilter } from "./table-search-filter"
-export type { TableSearchFilterProps } from "./table-search-filter"
-export { TableViewMenu } from "./table-view-menu"
-export type { TableViewMenuProps } from "./table-view-menu"
-export { TableClearFilter } from "./table-clear-filter"
-export type { TableClearFilterProps } from "./table-clear-filter"
-
-// Advanced filter components
-export { TableDateFilter } from "./table-date-filter"
-export { TableFacetedFilter } from "./table-faceted-filter"
-export { TableFilterMenu } from "./table-filter-menu"
-export { TableRangeFilter } from "./table-range-filter"
-export { TableSliderFilter } from "./table-slider-filter"
-
-// Advanced Rules Type Filter components
-export { TableSortMenu } from "./table-sort-menu"
-
-// Navigation components
-export { TablePagination } from "./table-pagination"
-export type { TablePaginationProps } from "./table-pagination"
-
-// Export components
-export { TableExportButton, exportTableToCSV } from "./table-export-button"
-export type {
-  TableExportButtonProps,
-  ExportTableToCSVOptions,
-} from "./table-export-button"
-
-// Column-level filter components
+// Core column-level filter components
 export { TableColumnTitle } from "./table-column-title"
 export { TableColumnActions } from "./table-column-actions"
-export {
-  TableColumnSortOptions,
-  TableColumnSortMenu,
-} from "./table-column-sort"
-export {
-  TableColumnHideOptions,
-  TableColumnHideMenu,
-} from "./table-column-hide"
-export { TableColumnPinOptions, TableColumnPinMenu } from "./table-column-pin"
-export {
-  TableColumnFacetedFilterOptions,
-  TableColumnFacetedFilterMenu,
-  TableColumnFilterTrigger,
-} from "./table-column-faceted-filter"
-export { TableColumnSliderFilterOptions } from "./table-column-slider-filter"
-export { TableColumnDateFilterOptions } from "./table-column-date-filter"
-
-// Row DnD components (standalone)
-export {
-  TableRowDndProvider,
-  TableDraggableRow,
-  TableRowDragHandle,
-} from "./table-row-dnd"
-export type {
-  TableRowDndProviderProps,
-  TableDraggableRowProps,
-  TableRowDragHandleProps,
-} from "./table-row-dnd"
-
-// Column DnD components (standalone)
-export {
-  TableColumnDndProvider,
-  TableDraggableHeader,
-  TableDragAlongCell,
-} from "./table-column-dnd"
-export type {
-  TableColumnDndProviderProps,
-  TableDraggableHeaderProps,
-  TableDragAlongCellProps,
-} from "./table-column-dnd"
