+++ src/registry/new-york/examples/niko-table/column-dnd-state.tsx
@@ -11,12 +11,12 @@ import {
   DataTableDndColumnBody,
 } from "@/components/niko-table/core/data-table-dnd-structure"
 import {
-  DataTableColumnDndProvider,
   DataTableEmptyIcon,
   DataTableEmptyMessage,
   DataTableEmptyTitle,
   DataTableEmptyDescription,
 } from "@/components/niko-table/components"
+import { DataTableColumnDndProvider } from "@/components/niko-table/components/data-table-column-dnd"
 import type { DataTableColumnDef } from "@/components/niko-table/types"
 import { Badge } from "@/components/ui/badge"
 import { Button } from "@/components/ui/button"
