+++ src/registry/new-york/examples/niko-table/row-dnd.tsx
@@ -9,8 +9,6 @@ import {
 } from "@/components/niko-table/core"
 import { DataTableDndBody } from "@/components/niko-table/core/data-table-dnd-structure"
 import {
-  DataTableRowDndProvider,
-  DataTableRowDragHandle,
   DataTableEmptyIcon,
   DataTableEmptyMessage,
   DataTableEmptyTitle,
@@ -18,6 +16,10 @@ import {
   DataTableColumnHeader,
   DataTableColumnTitle,
 } from "@/components/niko-table/components"
+import {
+  DataTableRowDndProvider,
+  DataTableRowDragHandle,
+} from "@/components/niko-table/components/data-table-row-dnd"
 import type { DataTableColumnDef } from "@/components/niko-table/types"
 import { Badge } from "@/components/ui/badge"
 import { Inbox } from "lucide-react"
