+++ src/registry/new-york/examples/niko-table/virtualized-column-dnd.tsx
@@ -8,12 +8,12 @@ import {
   DataTableVirtualizedDndColumnBody,
 } from "@/components/niko-table/core/data-table-virtualized-dnd-structure"
 import {
-  DataTableColumnDndProvider,
   DataTableEmptyIcon,
   DataTableEmptyMessage,
   DataTableEmptyTitle,
   DataTableEmptyDescription,
 } from "@/components/niko-table/components"
+import { DataTableColumnDndProvider } from "@/components/niko-table/components/data-table-column-dnd"
 import type { DataTableColumnDef } from "@/components/niko-table/types"
 import { Badge } from "@/components/ui/badge"
 import { Inbox } from "lucide-react"
